audio-inflator-alternative-2 - working version
With the audio on an empty and represented with a FakeEntity in A-Frame the sound works.
