+++ examples/playground/index.html
@@ -10,6 +10,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.5.0/socket.io.slim.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/open-easyrtc@2.0.20/api/easyrtc.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/networked-aframe@^0.12.0/dist/networked-aframe.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/gh/networked-aframe/naf-janus-adapter@3.2.0/examples/js/audio-system.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@fern-solutions/aframe-mirror@1.0.3/dist/mirror.umd.min.js"></script>
     <script src="https://cdn.jsdelivr.net/gh/c-frame/aframe-cursor-teleport@1.5.0/dist/aframe-cursor-teleport-component.min.js"></script>
     <script src="https://cdn.jsdelivr.net/gh/AdaRoseCannon/aframe-xr-boilerplate@f34468b9503d0c7326b9af0f1f09959004916875/simple-navmesh-constraint.js"></script>
@@ -17,14 +18,6 @@
     <script src="https://cdn.jsdelivr.net/npm/aframe-look-at-component@1.0.0/dist/aframe-look-at-component.min.js"></script>
     <script src="https://cdn.jsdelivr.net/gh/networked-aframe/naf-valid-avatars@4b116a4/src/player-info.js"></script>
     <script>
-      window.userClicked = false;
-
-      window.userHasClicked = () => {
-        if (!window.userClicked) {
-          window.userClicked = true;
-        }
-      };
-
       // @var url {string} The url as string.
       window.absoluteURLForAsset = (url) => {
         let newurl = "";
@@ -35,41 +28,6 @@
         }
 
         return newurl;
-      };
-
-      window.emitClick = () => {
-        // Select the button element
-        var button = document.querySelector('#btn-start');
-
-        // Create the click event
-        var clickEvent = new MouseEvent('click', {
-            'view': window,
-            'bubbles': true,
-            'cancelable': false
-        });
-
-        // Dispatch the event on the button
-        button.dispatchEvent(clickEvent);
-      }
-
-      window.playAllSounds = () => {
-        const sounds = document.querySelectorAll("[sound]");
-
-        if (sounds.length) {
-          for (const sound of sounds) {
-            sound.components.sound.playSound();
-          }
-        }
-      };
-
-      window.stopAllSounds = () => {
-        const sounds = document.querySelectorAll("[sound]");
-
-        if (sounds.length) {
-          for (const sound of sounds) {
-            sound.components.sound.stopSound();
-          }
-        }
       };
     </script>
     <script src="../../dist/gltf-model-plus.min.js"></script>
@@ -111,11 +69,6 @@
         },
       });
     </script>
-    <style>
-      .hidden {
-        display: none;
-      }
-    </style>
   </head>
   <body>
     <template id="avatar-template">
@@ -166,6 +119,8 @@
           <li>[g] toggle fly mode</li>
           <li>[shift down] move fast</li>
           <li>[space] toggle visibility of clickable waypoints</li>
+          <li>&nbsp;</li>
+          <li>[click on the screen once] start the sound</li>
         </ul>
       </div>
     </div>
@@ -188,6 +143,7 @@
     </script>
 
     <a-scene
+      audio
       light="defaultLightsEnabled:false"
       renderer="physicallyCorrectLights:true;colorManagement:true"
       networked-scene="
@@ -256,49 +212,12 @@
       </button>
     </div>
 
-    <div
-      id="confirm-start"
-      class=""
-      style="position: absolute; top: 0; left: 0; width: 100vw; height: 100vh; background-color: #ffffff99"
-    >
-      <button
-        id="btn-start"
-        class="hidden"
-        style="position: absolute; top: 0; left: 0; width: 100vw; height: 100vh"
-        onclick="javascript:this.parentNode.classList.add('hidden');window.userHasClicked();window.playAllSounds();"
-      >
-        <span
-          style="
-            margin-top: 50%;
-            margin-left: 0;
-            padding: 2rem;
-            background: rgba(0, 0, 0, 0.3);
-            border-radius: 0.5rem;
-            color: #ffffff;
-          "
-          >Please click to start</span
-        >
-      </button>
-    </div>
-
     <div class="naf-bottom-bar-center" style="display: none" id="uiInRoom">
       <naf-mic-button entity="#rig"></naf-mic-button>
       <naf-users-button></naf-users-button>
       <naf-chat-button></naf-chat-button>
     </div>
     <script>
-      // Check if the sceen has loaded to enable the start button.
-      // const scene = document.querySelector("a-scene");
-      // scene.addEventListener("renderstart", () => {
-      //   document.querySelector("#btn-start").classList.remove("hidden");
-      // });
-
-      const gltfModelPlus = document.querySelector("#gltf-model-plus-entity");
-      gltfModelPlus.addEventListener("model-loaded", () => {
-        console.log("- - - > model-loaded");
-        document.querySelector("#btn-start").classList.remove("hidden");
-      });
-
       // Add the drag and drop event listeners to the document and process a dragged GLB file.
       document.addEventListener(
         "dragover",
@@ -324,24 +243,10 @@
               const reader = new FileReader();
 
               reader.onload = function (e) {
-                // Show the confirm start dialogue if present.
-                const confirmStart = document.querySelector("#confirm-start");
-                if (confirmStart) {
-                  // window.createAudioContext();
-                  document.querySelector("#btn-start").classList.add("hidden");
-                  confirmStart.classList.remove("hidden");
-
-                  if (window.userClicked) {
-                    window.setTimeout(window.emitClick, 1000);
-                  }
-                }
-
-                // Stop all sound if there are sounds.
-                window.stopAllSounds();
-
                 // Set the gltf-model-plus attribute of the gltf-model-plus component to the dropped file.
                 // Set an empty string as source as a workaround to force the "model-loaded" event
                 // when the same GLB is loaded again.
+                const gltfModelPlus = document.querySelector("[gltf-model-plus]");
                 gltfModelPlus.setAttribute("gltf-model-plus", "");
                 gltfModelPlus.setAttribute("gltf-model-plus", e.target.result);
               };
